feat: :sparkles: Cria a rota para o endpoint /cars/id para excluir os registros de um carro

+++ src/models/MongoModel.ts
@@ -28,6 +28,7 @@ abstract class MongoModel<T> implements IModel<T> {
   }
 
   public async delete(_id:string):Promise<T | null> {
+    if (!isValidObjectId(_id)) throw new Error(ErrorTypes.InvalidMongoId);
     return this._model.findOneAndDelete({ _id });
   }
 }
